Move GetTitleAndBodyFromFlags to commands/utils
